Add getting started and development instructions to the README (#378)

+++ db/data/20221005110149_cleanup_to_avoid_carefully_migrating.rb
@@ -4,16 +4,16 @@ class CleanupToAvoidCarefullyMigrating < ActiveRecord::Migration[7.0]
 
     # there are no active users at this point so this is safe
     ActiveRecord::Base.transaction do
-      Releases::CommitListener.delete_all
-      BuildArtifact.delete_all
-      Releases::Step::Run.delete_all
-      SignOff.delete_all
-      Releases::Commit.delete_all
-      Releases::Step::Run.delete_all
-      Releases::Step.delete_all
-      Releases::Train::Run.delete_all
-      TrainSignOffGroup.delete_all
-      Releases::Train.delete_all
+      Releases::CommitListener.delete_all if defined?(Releases::CommitListener)
+      BuildArtifact.delete_all if defined?(BuildArtifact)
+      Releases::Step::Run.delete_all if defined?(Releases::Step::Run)
+      SignOff.delete_all if defined?(SignOff)
+      Releases::Commit.delete_all if defined?(Releases::Commit)
+      Releases::Step::Run.delete_all if defined?(Releases::Step::Run)
+      Releases::Step.delete_all if defined?(Releases::Step)
+      Releases::Train::Run.delete_all if defined?(Releases::Train::Run)
+      TrainSignOffGroup.delete_all if defined?(TrainSignOffGroup)
+      Releases::Train.delete_all if defined?(Releases::Train)
     end
   end
